Added documentation to PopupWindow
